Update document scripts to fix convertion issues and PHP warnings - refs BT#8274

+++ main/document/document.php
@@ -62,8 +62,10 @@ if(isset($_SESSION['temp_audio_nanogong']) && !empty($_SESSION['temp_audio_nanog
 	unlink($_SESSION['temp_audio_nanogong']);
 }
 
-if(isset($_SESSION['temp_realpath_image']) && !empty($_SESSION['temp_realpath_image'])) {
-	unlink($_SESSION['temp_realpath_image']);
+if(isset($_SESSION['temp_realpath_image']) && !empty($_SESSION['temp_realpath_image']) &&
+    file_exists($_SESSION['temp_realpath_image'])
+    ) {
+    unlink($_SESSION['temp_realpath_image']);
 }
 $courseInfo = api_get_course_info();
 $course_dir = $courseInfo['directory'].'/document';
@@ -261,9 +263,11 @@ switch ($action) {
         );
         $file = $sys_course_path.$courseInfo['directory'].'/document'.$document_info['path'];
         $fileInfo = pathinfo($file);
-        if (!(in_array($fileInfo['extension'], DocumentManager::get_jodconverter_extension_list('from', $formatType))) ||
+        if ($fileInfo['extension'] == $formatTarget) {
+            $message = Display::return_message(get_lang('ErrorSameFormat'), 'warning');
+        } elseif (!(in_array($fileInfo['extension'], DocumentManager::get_jodconverter_extension_list('from', $formatType))) ||
             !(in_array($formatTarget, DocumentManager::get_jodconverter_extension_list('to', $formatType)))) {
-            $message = Display::return_message(get_lang('FormatNotSupported'), 'error');
+            $message = Display::return_message(get_lang('FormatNotSupported'), 'warning');
         } else {
             $convertedFile = $fileInfo['dirname'] . DIRECTORY_SEPARATOR . $fileInfo['filename'] .
                 '_from_' . $fileInfo['extension'] . '.' . $formatTarget;
@@ -273,7 +277,7 @@ switch ($action) {
                 $message = Display::return_message(get_lang('FileExists'), 'error');
             } else {
                 $result = $obj->convertCopyDocument($file, $convertedFile, $convertedTitle);
-                if ((empty($result))) {
+                if (empty($result)) {
                     $message = Display::return_message(get_lang('CopyFailed'), 'error');
                 } else {
                     $cidReq = Security::remove_XSS($_GET['cidReq']);
@@ -289,7 +293,6 @@ switch ($action) {
                 }
             }
         }
-        Session::write('message', $message);
         break;
 }
